fixed the boot logic
